Refactor to separate the main feed and form components from the App component
